requirements.txt
langchain
langchain_community
langchain_google_genai
langchain_text_splitters
faiss-cpu
python-dotenv
streamlit
tiktoken
beautifulsoup4
SpeechRecognition
PyAudio